// src/enums/activitylogs.ts
export enum METHODS {
  CREATE = "CREATE",
  READ = "READ",
  UPDATE = "UPDATE",
  DELETE = "DELETE",
  PRINT = "PRINT",
  LOGIN = "LOGIN",
  LOGOUT = "LOGOUT",
  OPEN = "OPEN",
  SEND = "SEND",
  DOWNLOAD = "DOWNLOAD",
  PAY = "PAY",
}

export enum MODULES {
  ITEMCLASSIFICATION = "ITEM CLASSIFICATIONS",
  CURRENCY = "CURRENCY",
  CUSTOMER = "CUSTOMER",
  CATEGORY = "CATEGORY",
  BRANCH = "BRANCH",
  AREA = "AREA",
  PRINTERSTATION = "PRINTER STATION",
  ITEMSUBCLASSIFICATIONS = "ITEM SUB CLASSIFICATIONS",
  ITEMS = "ITEMS",
  TABLES = "TABLES",
  COMPANY = "COMPANY",
  SPECIALREQUEST = "SPECIAL REQUEST",
  VOIDREASONS = "VOID REASONS",
  DISCOUNTS = "DISCOUNTS",
  DINETYPE = "DINE TYPE",
  WAREHOUSE = "WAREHOUSE",
  CARDTYPES = "CARD TYPES",
  MEMC = "MEMC",
  PHYSICALCOUNTS = "PHYSICAL COUNTS",
  FREEREASONS = "FREE REASONS",
  CASHIOREASON = "CASH IN/OUT REASON",
  USERS = "USERS",
  SYSTEMPARAMETERS = "SYSTEM PARAMETERS",
  PAYMENTTYPES = "PAYMENT TYPES",
  TERMINALS = "TERMINALS",
  MALLHOOKUP = "MALL HOOK UP",
  HOME = "HOME",
  INVENTORYTRANSACTIONS = "INVENTORY TRANSACTIONS",
  HEADER = "HEADER",
  LOGIN = "LOGIN",
  LOGOUT = "LOGOUT",
  TABLEVIEW = "TABLE VIEW",
  CASHIN = "CASH IN",
  CASHOUT = "CASH OUT",
  CASHFUND = "CASH FUND",
  ORDERING = "ORDERING",
  ORDERPREVIEW = "ORDER PREVIEW",
  OTHERCHARGES = "OTHER CHARGES",
  PRICELIST = "PRICELIST",
  PRICEDETAILS = "PRICE DETAILS",
  CASHDECLARATION = "CASH DECLARATION",
  XREADING = "X-READING",
  ZREADING = "Z-READING",
  MANAGERSREPORT = "MANAGER'S REPORT",
  ORDERDETAILS = "ORDER DETAILS",
  DENOMINATION = "DENOMINATION",
  FOOTER = "FOOTER",
  IMPORT = "IMPORT",
  EXPORT = "EXPORT",
  REGENERATE_MALL_FILES = "REGENERATE MALL FILES",
  RECOMPUTE_ZREADING = "RECOMPUTE Z-READING",
  CHANGE_PASSWORD = "CHANGE PASSWORD",
  BACKUP_DATA = "BACKUP DATA",
  AUDIT_TRAIL = "AUDIT TRAIL",
  SYSTEM_SETTINGS = "SYSTEM SETTINGS",
}